requirements.txt
Flask==2.3.2
Flask-Login==0.6.2
mysql-connector-python==8.0.33
python-dotenv==1.0.0
matplotlib==3.7.2
boto3==1.26.162
Werkzeug==2.3.7